tests/test_trading_bot.py
"""
Integration tests for trading bot functionality.
These tests focus on trading bot use cases and strategies.
"""

from datetime import datetime, timedelta
from unittest.mock import MagicMock, Mock, patch

import pytest

import robin_stocks.gemini as gem
import robin_stocks.robinhood as rh
import robin_stocks.tda as tda


class TestTradingBotIntegration:
    """Test trading bot specific functionality."""

    @patch("robin_stocks.robinhood.helper.request_get")
    def test_portfolio_balance_check(self, mock_request):
        """Test checking portfolio balance before trading."""
        mock_request.return_value = {
            "total_return_today": "125.50",
            "total_return_today_percent": "0.0125",
            "market_value": "10000.00",
        }

        portfolio = rh.get_portfolio()
        balance = float(portfolio["market_value"])

        assert balance >= 0
        assert isinstance(balance, float)

    @patch("robin_stocks.robinhood.helper.request_get")
    def test_risk_management_position_size(self, mock_request):
        """Test position sizing for risk management."""
        # Mock portfolio value
        mock_request.return_value = {"market_value": "10000.00"}

        portfolio_value = float(rh.get_portfolio()["market_value"])
        max_position_percent = 0.05  # 5% max per position
        max_position_value = portfolio_value * max_position_percent

        assert max_position_value == 500.0
        assert max_position_value < portfolio_value

    @patch("robin_stocks.robinhood.helper.request_get")
    def test_stop_loss_calculation(self, mock_request):
        """Test stop loss calculation."""
        mock_request.return_value = [{"price": "150.00"}]

        current_price = float(rh.get_latest_price("AAPL")[0])
        stop_loss_percent = 0.02  # 2% stop loss
        stop_loss_price = current_price * (1 - stop_loss_percent)

        assert stop_loss_price == 147.0
        assert stop_loss_price < current_price


class TestMarketDataAnalysis:
    """Test market data analysis for trading decisions."""

    @patch("robin_stocks.robinhood.helper.request_get")
    def test_moving_average_calculation(self, mock_request):
        """Test calculating moving averages from price data."""
        # Mock historical price data
        mock_request.return_value = [
            {"close_price": "100.00"},
            {"close_price": "102.00"},
            {"close_price": "101.00"},
            {"close_price": "103.00"},
            {"close_price": "105.00"},
        ]

        prices = [float(p["close_price"]) for p in mock_request.return_value]
        moving_avg = sum(prices) / len(prices)

        assert moving_avg == 102.2
        assert len(prices) == 5

    @patch("robin_stocks.gemini.helper.request_get")
    def test_crypto_volatility_check(self, mock_request):
        """Test checking crypto volatility for trading decisions."""
        mock_request.return_value = {
            "bid": "45000.00",
            "ask": "45500.00",
            "last": "45250.00",
        }

        ticker_data = gem.get_pubticker("btcusd")
        spread = float(ticker_data["ask"]) - float(ticker_data["bid"])
        spread_percent = spread / float(ticker_data["last"])

        assert spread == 500.0
        assert spread_percent > 0


class TestOrderManagement:
    """Test order management functionality."""

    def test_order_validation(self):
        """Test order parameter validation."""
        # Test valid order parameters
        symbol = "AAPL"
        quantity = 10
        order_type = "market"
        side = "buy"

        assert symbol.isalpha()
        assert quantity > 0
        assert order_type in ["market", "limit", "stop"]
        assert side in ["buy", "sell"]

    @patch("robin_stocks.robinhood.helper.request_post")
    def test_dry_run_order(self, mock_request):
        """Test dry run (paper trading) functionality."""
        mock_request.return_value = {
            "id": "test-order-123",
            "state": "queued",
            "price": "150.00",
            "quantity": "10",
        }

        # Simulate dry run by not actually placing order
        dry_run = True
        if not dry_run:
            order = rh.orders.order_buy_market("AAPL", 10)
        else:
            # Just validate parameters
            order = {"status": "dry_run_success", "symbol": "AAPL", "quantity": 10}

        assert order["status"] == "dry_run_success"


class TestMultiApiStrategy:
    """Test strategies using multiple APIs."""

    @patch("robin_stocks.robinhood.helper.request_get")
    @patch("robin_stocks.tda.helper.request_get")
    def test_price_comparison_arbitrage(self, mock_tda, mock_rh):
        """Test comparing prices across different platforms."""
        # Mock different prices on different platforms
        mock_rh.return_value = [{"price": "150.00"}]
        mock_tda_response = Mock()
        mock_tda_response.json.return_value = {"AAPL": {"lastPrice": 150.50}}
        mock_tda_response.status_code = 200
        mock_tda.return_value = (mock_tda_response, None)

        rh_price = float(rh.get_latest_price("AAPL")[0])
        tda_response, _ = tda.get_quote("AAPL")
        tda_price = tda_response.json()["AAPL"]["lastPrice"]

        price_diff = abs(tda_price - rh_price)
        assert price_diff >= 0

    @patch("robin_stocks.gemini.helper.request_get")
    @patch("robin_stocks.robinhood.helper.request_get")
    def test_crypto_stock_correlation(self, mock_rh, mock_gem):
        """Test analyzing correlation between crypto and stock markets."""
        # Mock crypto price
        mock_gem.return_value = {"last": "45000.00"}

        # Mock stock price
        mock_rh.return_value = [{"price": "150.00"}]

        crypto_price = float(gem.get_pubticker("btcusd")["last"])
        stock_price = float(rh.get_latest_price("AAPL")[0])

        # Simple correlation test (in real bot, use proper correlation calculation)
        assert crypto_price > 0
        assert stock_price > 0


class TestBacktesting:
    """Test backtesting functionality."""

    def test_backtest_strategy_setup(self):
        """Test setting up a backtesting environment."""
        # Strategy parameters
        strategy = {
            "name": "moving_average_crossover",
            "parameters": {
                "short_ma": 10,
                "long_ma": 20,
                "stop_loss": 0.02,
                "take_profit": 0.05,
            },
        }

        assert strategy["parameters"]["short_ma"] < strategy["parameters"]["long_ma"]
        assert strategy["parameters"]["stop_loss"] > 0
        assert strategy["parameters"]["take_profit"] > 0

    def test_performance_metrics(self):
        """Test calculating performance metrics."""
        # Mock trading results
        trades = [
            {"entry": 100, "exit": 105, "profit": 5},
            {"entry": 105, "exit": 103, "profit": -2},
            {"entry": 103, "exit": 108, "profit": 5},
        ]

        total_profit = sum(trade["profit"] for trade in trades)
        win_rate = len([t for t in trades if t["profit"] > 0]) / len(trades)

        assert total_profit == 8
        assert win_rate == 2 / 3


class TestRiskManagement:
    """Test risk management features."""

    def test_position_size_calculation(self):
        """Test Kelly criterion for position sizing."""
        portfolio_value = 10000
        win_rate = 0.6
        avg_win = 5
        avg_loss = 3

        # Simplified Kelly criterion
        kelly_percent = (win_rate * avg_win - (1 - win_rate) * avg_loss) / avg_win
        kelly_percent = max(0, min(0.25, kelly_percent))  # Cap at 25%

        position_size = portfolio_value * kelly_percent

        assert position_size >= 0
        assert position_size <= portfolio_value * 0.25

    def test_daily_loss_limit(self):
        """Test daily loss limit enforcement."""
        starting_balance = 10000
        current_balance = 9800
        daily_loss_limit = 0.02  # 2%

        daily_loss = (starting_balance - current_balance) / starting_balance

        assert daily_loss == 0.02

        # Should stop trading if limit reached
        should_stop_trading = daily_loss >= daily_loss_limit
        assert should_stop_trading


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
